chore(clean): cleaning the codebase for submission
