+++ server.js
@@ -3,20 +3,16 @@ const http = require('http');
 const { Server } = require('socket.io');
 const app = require('./src/app');
 const { testConnection } = require('./src/utils/db');
-const { checkAndRestoreDatabase } = require('./src/utils/db-backup');
 const setupGameSocket = require('./src/sockets/game-socket'); // Assuming this is for game logic
-const { router, rooms } = require('./src/routes/game-routes');
+const { rooms } = require('./src/routes/game-routes');
 const server = http.createServer(app);
 const io = new Server(server);
 
 const PORT = process.env.PORT || 3000;
 
-// Check database connection and restore if necessary
-checkAndRestoreDatabase()
+// Test database connection
+testConnection()
   .then(() => {
-    // Test database connection
-    testConnection();
-
     // Initialize Socket.IO once the database is ready
     setupGameSocket(io, rooms);  // Assuming setupGameSocket sets up the necessary game event listeners
