+++ src/app.js
@@ -4,7 +4,6 @@ const express = require('express');
 const app = express();
 const { router } = require('./routes/user-routes.js');
 const { router: gameRouter } = require('./routes/game-routes.js');
-const voteRouter = require('./routes/vote-routes.js');
 const { router : adminRouter} = require("./routes/admin-routes.js");
 
 // Middleware for JSON and URL-encoded body parsing
@@ -17,7 +16,6 @@ app.use(express.static(path.join(__dirname, 'public')));
 // API routes
 app.use('/api/users', router);
 app.use('/api/game', gameRouter);
-app.use('/api/votes', voteRouter);
 app.use('/api/admin', adminRouter);
 
 // 404 Handler for any undefined routes
