--- src/public/game/voting-round.html
@@ -1,40 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Voting Round</title>
-    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.0/dist/css/bootstrap.min.css" rel="stylesheet">
-</head>
-
-<body class="bg-light">
-    <div class="container py-5">
-        <h1 class="text-center mb-4">Voting Round</h1>
-        <div class="card shadow-sm p-4">
-            <form id="vote-form">
-                <div class="mb-3">
-                    <label class="form-label">Current Voter:</label>
-                    <h4 id="current-voter" class="text-primary"></h4>
-                </div>
-                <div class="mb-3">
-                    <label for="vote-for" class="form-label">Vote For:</label>
-                    <select id="vote-for" class="form-select" required></select>
-                </div>
-                <button type="submit" class="btn btn-primary w-100">Cast Vote</button>
-            </form>
-            <div id="results" class="mt-4"></div>
-            <button id="vote-again-btn" class="btn btn-secondary mt-3 d-none">Vote Again</button>
-        </div>
-
-        <div class="card shadow-sm p-4 mt-4">
-            <h5 class="card-title">Vote Counts</h5>
-            <div id="vote-counts"></div>
-        </div>
-    </div>
-    
-    <script src="/socket.io/socket.io.js"></script>
-    <script src="/js/game/voting-round.js"></script>
-</body>
-
-</html>
