+++ src/public/js/utils/game-utils.js
@@ -1,4 +1,3 @@
-
 /**
  * 
  * @param Array<Object> players
